+++ src/tests/pages/ComplaintDetails.test.tsx
@@ -8,6 +8,8 @@ import {
 import userEvent from '@testing-library/user-event';
 import ComplaintDetails from '../../pages/ComplaintDetails';
 import api from '../../services/api';
+import { MemoryRouter, Router } from 'react-router-dom';
+import { createMemoryHistory } from 'history';
 
 const mockHistoryPush = jest.fn();
 const mockReactRouterDom = jest.fn();
@@ -21,6 +23,27 @@ jest.mock('react-router-dom', () => ({
 		id: 1,
 	}),
 }));
+
+function mockComplaint(open: boolean) {
+	return {
+		data: {
+			complaint_id: 1,
+			complaint_name: 'Geno',
+			complaint_description: 'Thyroid vessel ligation',
+			complaint_latitude: 31.975314,
+			complaint_longitude: 35.196042,
+			complaint_userId: 1,
+			complaint_category: 'Hole',
+			complaint_creationDate: '2021-03-12T10:02:31.000Z',
+			complaint_closeDate: '2021-11-25T19:05:09.000Z',
+			complaint_status: open ? 'open' : 'wait',
+			vote_id: null,
+			vote_userId: null,
+			vote_complaintId: null,
+			vote_typeVote: null,
+		},
+	};
+}
 
 const complaintWithVoteMock = (status: string, hasVote = true) => ({
 	complaint_id: 1,
@@ -39,6 +62,25 @@ const complaintWithVoteMock = (status: string, hasVote = true) => ({
 	vote_userId: hasVote ? 1 : null,
 	vote_complaintId: hasVote ? 7 : null,
 	vote_typeVote: hasVote ? 'complaintUpvote' : null,
+});
+
+const complaintWithVoteMockDelete = (status: string) => ({
+	complaint_id: 1,
+	complaint_name: 'Sub-Ex',
+	complaint_description:
+		'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.',
+	complaint_latitude: 36.275231,
+	complaint_longitude: 113.310158,
+	complaint_userId: 1,
+	complaint_category: 'Water',
+	complaint_creationDate: '2021-02-21T18:52:45.000Z',
+	complaint_closeDate: '2020-11-11T05:41:31.000Z',
+	complaint_picture: 'http://dummyimage.com/237x100.png/cc0000/ffffff',
+	complaint_status: status,
+	vote_id: 405,
+	vote_userId: 1,
+	vote_complaintId: 7,
+	vote_typeVote: 'complaintUpvote',
 });
 
 describe('Test ComplaintDetails With waiting complaint with vote', () => {
@@ -116,5 +158,46 @@ describe('Test ComplaintDetails With open complaint with no vote', () => {
 
 		fireEvent.click(confirmButton);
 		expect(screen.getByText('DENÚNCIA REPORTADA')).toBeInTheDocument();
+	});
+});
+
+describe('Teste ComplaintDetails with open complaint with confirm button', () => {
+	test('test confirm complaint click', async () => {
+		jest.spyOn(api, 'get').mockImplementationOnce(() =>
+			Promise.resolve(mockComplaint(false)),
+		);
+		jest.spyOn(api, 'post').mockImplementationOnce(() => Promise.resolve());
+		render(<ComplaintDetails />);
+
+		await waitFor(() => screen.getByText('Geno'));
+		const confirmButton = screen.getByTestId('button');
+
+		fireEvent.click(confirmButton);
+		expect(screen.getByText('PROBLEMA RESOLVIDO')).toBeInTheDocument();
+	});
+});
+
+describe('Test ComplaintDetails', () => {
+	test('test delete complaint click', async () => {
+		jest.spyOn(api, 'get').mockImplementationOnce(() =>
+			Promise.resolve({ data: complaintWithVoteMockDelete('open') }),
+		);
+		const mockApi = jest.fn();
+		jest.spyOn(api, 'delete').mockImplementationOnce(mockApi);
+
+		act(() => {
+			render(
+				<MemoryRouter>
+					<Router history={createMemoryHistory()}>
+						<ComplaintDetails />
+					</Router>
+				</MemoryRouter>,
+			);
+		});
+
+		await waitFor(() => screen.getByText('Sub-Ex'));
+
+		userEvent.click(screen.getByText('Deletar Denúncia'));
+		expect(mockApi).toHaveBeenCalled();
 	});
 });
